Day 15
Day 15 started with installing python and pycharm on my machine and starting to work from that instead of using Repl.it. The challenge at the end was to write the code for a coffee machine! Very appropriate for me! It only took me a couple of hours to get this up and running. I'm sure professionals could do this a lot faster in their sleep, but I'm still chuffed with myself for figuring it all out!